requirements.txt
langchain
langchain-community==0.2.12
langchain-pinecone==0.1.3
flask
python-dotenv
PyPDF2 
pdfplumber
-e .

